requirements.txt
Django==3.1
pip==20.2.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.4.0
django-filter==2.3.0